requirements.txt
langchain_openai
langchain_core
langchainhub
langchain_groq
python-dotenv
streamlit
altair
langchain_community
langserve
fastapi
uvicorn
sse_starlette
ipykernel
bs4
pypdf
chromadb
faiss-cpu
arxiv
wikipedia